Move overlay content key types back into trin-core
