+++ newsfragments/276.fixed.md
@@ -1,0 +1,1 @@
+Move overlay content key types back into trin-core.
