+++ trin-core/Cargo.toml
@@ -38,6 +38,7 @@ r2d2_sqlite = "0.19.0"
 serde = {version = "1.0.125", features = ["derive"] }
 serde_json = "1.0.59"
 serial_test = "0.5.1"
+sha2 = "0.10.1"
 stunclient = "0.1.2"
 structopt = "0.3"
 sha3 = "0.9.1"
