+++ trin-history/Cargo.toml
@@ -11,14 +11,8 @@ rust-version = "1.58.0"
 anyhow = "1.0.52"
 bytes = "1.1.0"
 discv5 = "0.1.0-beta.13"
-eth2_ssz = "0.4.0"
-eth2_ssz_derive = "0.3.0"
-eth2_ssz_types = "0.2.1"
-ethereum-types = "0.12.0"
 hex = "0.4.3"
 keccak-hash = "0.8.0"
-sha2 = "0.10.1"
-sha3 = "0.9.1"
 log = "0.4.14"
 rlp = "0.5.0"
 rocksdb = "0.16.0"
