+++ trin-history/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod content_key;
 pub mod events;
 mod jsonrpc;
 pub mod network;
