+++ trin-state/src/network.rs
@@ -10,6 +10,7 @@ use trin_core::portalnet::{
     overlay::{OverlayConfig, OverlayProtocol, OverlayRequestError},
     storage::{PortalStorage, PortalStorageConfig},
     types::{
+        content_key::StateContentKey,
         messages::{PortalnetConfig, ProtocolId},
         metric::XorMetric,
     },
@@ -17,7 +18,7 @@ use trin_core::portalnet::{
 use trin_core::utils::db::setup_overlay_db;
 use trin_core::utp::stream::UtpListener;
 
-use crate::{content_key::StateContentKey, trie::TrieDB};
+use crate::trie::TrieDB;
 
 /// State network layer on top of the overlay protocol. Encapsulates state network specific data and logic.
 #[derive(Clone)]
